refactor(misc): keyed phase_results + phase-owned tracking & reporting (#239)
