+++ src/raitap/pipeline/outputs.py
@@ -3,16 +3,34 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any, ClassVar, Protocol, runtime_checkable
 
 from raitap.types import TaskKind
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     import torch
 
-    from raitap.metrics import MetricsEvaluation
-    from raitap.robustness.results import RobustnessResult, RobustnessVisualisationResult
-    from raitap.transparency.results import ExplanationResult, VisualisationResult
+    from raitap.reporting.sections import ReportContext, ReportSection
+    from raitap.tracking.base_tracker import BaseTracker
+
+
+@runtime_checkable
+class PhaseResult(Protocol):
+    """An assessment phase's result: loggable to a tracker and renderable to a report.
+
+    Each phase (metrics / transparency / robustness / …) returns one of these
+    from its ``run()``; ``RunOutputs.phase_results`` keys them by phase name. The
+    pipeline iterates them generically — it never names a concrete phase — so a
+    new module plugs in without touching ``pipeline/`` or ``reporting/``.
+    """
+
+    report_order: ClassVar[int]
+
+    def log(self, tracker: BaseTracker | None, **kwargs: Any) -> None: ...
+
+    def report_sections(self, ctx: ReportContext) -> Sequence[ReportSection]: ...
 
 
 @dataclass(frozen=True)
@@ -54,14 +72,16 @@ class ForwardOutput:
 
 @dataclass(frozen=True)
 class RunOutputs:
-    """Assessment outputs; ``forward_output`` is the typed model forward output."""
+    """Assessment outputs.
 
-    explanations: list[ExplanationResult]
-    visualisations: list[VisualisationResult]
-    metrics: MetricsEvaluation | None
+    ``forward_output`` is the typed model forward output; ``phase_results`` holds
+    each configured assessment phase's result keyed by phase name (``"metrics"``,
+    ``"transparency"``, ``"robustness"``, …). Each value is a :class:`PhaseResult`
+    (``Trackable`` + ``Reportable``); the pipeline iterates them generically.
+    """
+
     forward_output: ForwardOutput
+    phase_results: dict[str, PhaseResult] = field(default_factory=dict)
     sample_ids: list[str] | None = None
     targets: torch.Tensor | list[dict[str, torch.Tensor]] | None = None
     prediction_summaries: tuple[PredictionSummary, ...] = ()
-    robustness_results: list[RobustnessResult] = field(default_factory=list)
-    robustness_visualisations: list[RobustnessVisualisationResult] = field(default_factory=list)
