+++ src/raitap/reporting/tests/test_formal_section.py
@@ -22,6 +22,7 @@ from raitap.robustness.contracts import (
     RobustnessVerdict,
     ThreatModel,
 )
+from raitap.robustness.report import RobustnessPhaseResult
 from raitap.robustness.results import RobustnessMetrics, RobustnessResult, encode_verdicts
 from raitap.types import TaskKind
 
@@ -82,13 +83,9 @@ def test_build_report_renders_robustness_certification_section(tmp_path: Path) -
     result = _formal_result(run_dir)
 
     outputs = RunOutputs(
-        explanations=[],
-        visualisations=[],
-        metrics=None,
         forward_output=_fo(torch.zeros(3, 5)),
         sample_ids=None,
-        robustness_results=[result],
-        robustness_visualisations=[],
+        phase_results={"robustness": RobustnessPhaseResult(robustness_results=[result])},
     )
 
     report = build_report(config, outputs)
@@ -131,13 +128,9 @@ def test_build_report_includes_per_class_bound_rows_for_formal_results(
     }
 
     outputs = RunOutputs(
-        explanations=[],
-        visualisations=[],
-        metrics=None,
         forward_output=_fo(torch.zeros(3, 5)),
         sample_ids=None,
-        robustness_results=[result],
-        robustness_visualisations=[],
+        phase_results={"robustness": RobustnessPhaseResult(robustness_results=[result])},
     )
 
     report = build_report(config, outputs)
@@ -177,13 +170,9 @@ def test_build_report_excludes_rows_with_only_lower_bound(tmp_path: Path) -> Non
     }
 
     outputs = RunOutputs(
-        explanations=[],
-        visualisations=[],
-        metrics=None,
         forward_output=_fo(torch.zeros(3, 5)),
         sample_ids=None,
-        robustness_results=[result],
-        robustness_visualisations=[],
+        phase_results={"robustness": RobustnessPhaseResult(robustness_results=[result])},
     )
 
     report = build_report(config, outputs)
@@ -217,13 +206,9 @@ def test_build_report_skips_bound_rows_for_empirical_attack_results(
     }
 
     outputs = RunOutputs(
-        explanations=[],
-        visualisations=[],
-        metrics=None,
         forward_output=_fo(torch.zeros(3, 5)),
         sample_ids=None,
-        robustness_results=[result],
-        robustness_visualisations=[],
+        phase_results={"robustness": RobustnessPhaseResult(robustness_results=[result])},
     )
 
     report = build_report(config, outputs)
